[#171953217] generateApi exported as function
