+++ e2e/README.md
@@ -2,12 +2,13 @@
 In this folder is defined a barebone project used to test `gen-api-models`' generated files in real-world scenarios. 
 
 ## Usage
-The entrypoint is the `run.sh` script which does the following:
-1. Compiles the `gen-api-models` program from source
-2. Execute `gen-api-models` against a test api specification
-3. Run tests against generated files
 
-### Examples
+```sh
+$ yarn install --frozen-lockfile
+$ yarn start
+```
+
+Please be sure that the `italia-utils` module has been compiled first. Otherwise you can use the `run.sh` script that does everything for you.
 
 ```sh
 $ cd .. # io-utils source root
