+++ e2e/package.json
@@ -3,14 +3,13 @@
   "version": "0.0.1",
   "license": "MIT",
   "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "tsc",
-    "test": "jest",
+    "start": "jest",
     "lint": "tslint --project ."
   },
   "devDependencies": {
     "@types/jest": "^24.0.13",
     "@types/node": "10.14.1",
+    "fs-extra": "^9.0.0",
     "jest": "^24.8.0",
     "rimraf": "^2.6.2",
     "ts-jest": "^24.0.2",
@@ -19,7 +18,8 @@
   "dependencies": {
     "fp-ts": "1.12.0",
     "io-ts": "1.8.5",
-    "italia-ts-commons": "^5.1.13"
+    "italia-ts-commons": "^5.1.13",
+    "italia-utils": "file:../"
   },
   "jest": {
     "testEnvironment": "node",
@@ -34,6 +34,8 @@
       "tsx"
     ],
     "preset": "ts-jest",
-    "testMatch": null
+    "testMatch": null,
+    "globalSetup": "./setup.ts",
+    "globalTeardown": "./teardown.ts"
   }
 }
