+++ e2e/src/__tests__/test-api/definitions.test.ts
@@ -1,5 +1,6 @@
+
 describe("Decoders generated from Test API spec defintions", () => {
-  const MODULE_PATH = `${process.cwd()}/generated/test-api`;
+  const MODULE_PATH = process.env.OUTPUT_DIR;
   const loadModule = (name: string) =>
     import(`${MODULE_PATH}/${name}.ts`).then(mod => {
       if (!mod) {
