+++ e2e/src/__tests__/test-api/request-types.test.ts
@@ -7,7 +7,7 @@ const mockResponse = (status: number, body?: any, headers?: any) => ({
 });
 
 describe("Request types generated from Test API spec", () => {
-  const MODULE_PATH = `${process.cwd()}/generated/test-api`;
+  const MODULE_PATH = process.env.OUTPUT_DIR;
   const loadModule = () =>
     import(`${MODULE_PATH}/requestTypes.ts`).then(mod => {
       if (!mod) {
