+++ e2e/teardown.ts
@@ -1,0 +1,7 @@
+import { remove } from "fs-extra";
+
+export default async () => {
+  if (process.env.OUTPUT_DIR) {
+    await remove(process.env.OUTPUT_DIR);
+  }
+};
