+++ e2e/tsconfig.json
@@ -11,6 +11,5 @@
     "sourceMap": true,
     "lib": ["es2017", "dom"]
   },
-  "include": ["src/**/*"],
   "exclude": ["node_modules", "src/**/__tests__/*"]
 }
