+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
-    "target": "es2018",
-    "lib": ["es2018"],
+    "target": "es6",
+    "lib": ["es6"],
     "rootDir": "src",
     "outDir": "dist",
     "module": "commonjs",
